# perf(react-router): remove `TRouterContext` for each route, lazily infer parent for merging and remove `TParentRoute` from options