--- a/packages/react-router/src/route.ts
+++ b/packages/react-router/src/route.ts
@@ -678,7 +678,6 @@ export class Route<
     children: TChildren
     loaderData: TLoaderData
     loaderDeps: TLoaderDeps
-    routerContext: any
   }
 
   init = (opts: { originalIndex: number }): void => {