From my understanding, `TRouterContext` is only relevant to the `RootRoute` and is not relevant to each route. If we remove it from the `Route` but keep it in `RootRoute` we can save TS some work. We can also remove `TParentRoute` from route options only for file based routing, as `TParentRoute` is a complex type which is inferred this can also save some work. We can use infer the parents `search`, `params` and `context` instead of indexing `TParentRoute`. 


Before
```
> tsc --extendedDiagnostics

Files:                         643
Lines of Library:            38411
Lines of Definitions:        83449
Lines of TypeScript:         19754
Lines of JavaScript:             0
Lines of JSON:                   0
Lines of Other:                  0
Identifiers:                135268
Symbols:                    327649
Types:                      115799
Instantiations:             814838
Memory used:               356762K
Assignability cache size:    48649
Identity cache size:          8548
Subtype cache size:           2426
Strict subtype cache size:     413
I/O Read time:               0.07s
Parse time:                  0.56s
ResolveModule time:          0.13s
ResolveTypeReference time:   0.00s
ResolveLibrary time:         0.02s
Program time:                0.87s
Bind time:                   0.26s
Check time:                  5.40s
printTime time:              0.00s
Emit time:                   0.00s
Total time:                  6.53s
```

After
```
> tsc --extendedDiagnostics

Files:                         643
Lines of Library:            38411
Lines of Definitions:        83463
Lines of TypeScript:         21604
Lines of JavaScript:             0
Lines of JSON:                   0
Lines of Other:                  0
Identifiers:                134834
Symbols:                    320440
Types:                      114112
Instantiations:             787854
Memory used:               356286K
Assignability cache size:    49024
Identity cache size:          8554
Subtype cache size:           2426
Strict subtype cache size:     413
I/O Read time:               0.09s
Parse time:                  0.61s
ResolveModule time:          0.13s
ResolveTypeReference time:   0.00s
ResolveLibrary time:         0.02s
Program time:                0.94s
Bind time:                   0.32s
Check time:                  5.27s
printTime time:              0.00s
Emit time:                   0.00s
Total time:                  6.53s
```